refactor(web-e2e): split browser scenario runner
